requirements.txt
flask
matplotlib==3.1.2
opencv-python-headless
torch==1.9.0
torchvision==0.10.0
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0